Consolidate types across Calendar and Availabilities
